requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
feedgen==1.0.0
pygments==2.17.2
